Implemented 1to1 VC joining button for trainee
